Decouple typescript language and grammar contribution
