+++ packages/textmate-grammars/data/snippets/typescript.json
