+++ packages/textmate-grammars/data/typescript.tmlanguage.json
