+++ packages/textmate-grammars/data/typescript.tsx.tmlanguage.json
