+++ packages/textmate-grammars/src/browser/textmate-grammars-frontend-module.ts
@@ -48,6 +48,9 @@ import { XmlContribution } from './xml';
 import { XslContribution } from './xsl';
 import { JavaContribution } from './java';
 import { LanguageGrammarDefinitionContribution } from '@theia/monaco/lib/browser/textmate/textmate-contribution';
+import { TypescriptContribution } from './ts';
+import { JavascriptContribution } from './js';
+import { JsxTagsContribution } from './jsx-tags';
 
 export default new ContainerModule(bind => {
     bind(BatContribution).toSelf().inSingletonScope();
@@ -145,4 +148,13 @@ export default new ContainerModule(bind => {
 
     bind(JavaContribution).toSelf().inSingletonScope();
     bind(LanguageGrammarDefinitionContribution).toService(JavaContribution);
+
+    bind(TypescriptContribution).toSelf().inSingletonScope();
+    bind(LanguageGrammarDefinitionContribution).toService(TypescriptContribution);
+
+    bind(JavascriptContribution).toSelf().inSingletonScope();
+    bind(LanguageGrammarDefinitionContribution).toService(JavascriptContribution);
+
+    bind(JsxTagsContribution).toSelf().inSingletonScope();
+    bind(LanguageGrammarDefinitionContribution).toService(JsxTagsContribution);
 });
